Add files via upload
Web Scrapping on Page with Pagination. Open a Terminal in the Jupyter notebook then create a project. After creating the project, you have to generate a new spider using pre-defined templates. After that, create a loop to extract the data that you need. Then obtain a pagination. When it's all done export it to a CSV file.