Create Scrolling
Web Scrapping on Page with Scrolling Open a Terminal in the Jupyter notebook then create a project. After creating the project, you have to generate a new spider using pre-defined templates. Import json.  After that, create a loop to extract the data that you need in the website. When it's all done export it to a CSV file.